Separate the request models from the response models
